requirements.txt
Django==3.1.0
djangorestframework==3.11.1

flake8==3.8.3